# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API de Pagamentos Sindireceita
  version: "1.1.0"
  summary: Cobrança e conciliação (PIX, boleto, checkout, DDA) sobre bancos do cliente.
  description: |
    API REST `/v1` que empresas usam como meio de cobrança sobre a **sua própria**
    credencial bancária (modelo PSP-Indireto — o dinheiro liquida direto na conta
    da empresa). Primeiro banco em produção: **C6**.

    Guia de integração/onboarding: `integration-guide.md` (mesma pasta).
    Manual de casos de uso end-to-end (cookbook com `curl` por caso):
    `manual.md` (mesma pasta) — cada caso referencia o `operationId` daqui.

    Além do plano `/v1` cliente-facing, este contrato documenta o **plano
    administrativo `/admin/*`** (tag `admin`, autenticado por token de admin —
    NÃO por token de empresa-cliente) e o **webhook inbound** de liquidação do
    C6 (tag `Webhooks`), do ponto de vista do integrador.

    Convenções globais:
    - **Autenticação:** `Authorization: Bearer <token>` em toda rota `/v1`. A
      empresa-cliente é derivada do token, nunca de um parâmetro. Token
      ausente/inválido → `401`.
      **Ressalva do caminho chave-de-Conta (revendedor, modelo b — ADR-0011):**
      quando o Bearer é uma **chave-de-Conta** (não um token de empresa-cliente),
      a empresa-cliente é escolhida por chamada no header `X-Client-Tenant`. Esse
      seletor é o **único mecanismo AUTORIZADO** de seleção, mediado pelo guard de
...
      do extrato que usa `YYYY-MM-DD`.
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/clients:
    post:
      operationId: provisionClient
      tags: [Revenda]
      summary: Provisionar empresa-cliente da Conta (modelo b, flag-gated)
      description: |
        **Modelo (b) — ADR-0011 §4. Flag-gated (`PAYMENT_ACCOUNT_KEY_SELECTOR`),
        default-off.** Autenticada **pela chave-de-Conta**, cria uma
        empresa-cliente (tenant) já vinculada à Conta e devolve o `tenant_id`
        para uso no seletor `X-Client-Tenant`.

        **Segurança (A01/T6):** o `account_id` da nova empresa-cliente vem
        **sempre da chave** (server-side, imutável — invariante set-once,
        ADR-0009 §2). O corpo **não** aceita `account_id`; enviá-lo é ignorado/
        rejeitado. Assim uma Conta nunca provisiona empresa-cliente sob outra
        Conta.

        A credencial bancária da nova empresa-cliente segue por
        `PUT /v1/bank-credential` self-serve (SIN-69196), agora endereçável pelo
        seletor. `Idempotency-Key` obrigatório (dedup em retry).

        **Webhook (SIN-69584 / B1):** este endpoint **não** cunha nem devolve
        `webhook_ref`/`webhook_path`. A **única** ref de callback C6 nasce no
        ponto de convergência in-flow (credencial + chave PIX completas,
        SIN-69560 / F2) e o texto-claro **nunca** sai do processo (só o hash é
        persistido). A Conta não precisa da ref — quem chama o webhook é o PSP
        C6, não a Conta. Uma empresa-cliente recém-criada tem **zero** refs até a
        convergência, e depois exatamente **uma**.
      security:
        - accountKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClientRequest" }
      responses:
        "201":
          description: Empresa-cliente criada e vinculada à Conta da chave.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token de API escopado por empresa-cliente (modelo a).
    accountKeyAuth:
      type: http
      scheme: bearer
      description: >-
        Chave-de-Conta (modelo b, ADR-0011, flag-gated default-off). Segredo
        opaco `ak_…` (≥256-bit) que identifica a **Conta** revendedora, não uma
        empresa-cliente. Exige o header `X-Client-Tenant` a cada chamada `/v1`
        para escolher a empresa-cliente-alvo (mediado pelo guard §2). Emitida/
        rotacionada via `POST /v1/account-key`.
    adminAuth:
      type: http
      scheme: bearer
      description: >-
        Token de **admin** do plano `/admin/*` (interno). Deny-by-default: um
        token de empresa-cliente ou uma chave-de-Conta nunca resolve papel de
        admin e é rejeitado. O papel `admin` autoriza mutações; `operator` é
        somente-leitura. Distribuído fora de banda pela Sindireceita; nunca
        em URL/log.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Chave estável por intenção de negócio. Reenvio com a mesma chave não duplica o efeito.
      schema: { type: string, minLength: 1 }
...
  responses:
    BadRequest:
      description: Validação falhou, JSON malformado, campo desconhecido ou header obrigatório ausente.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid: { value: { error: "invalid request" } }
            idem: { value: { error: "missing Idempotency-Key header" } }
    Unauthorized:
      description: Token ausente ou inválido.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "unauthorized" }
...
    TooManyRequests:
      description: Rate limit excedido (burst 20, 10 req/s por empresa). Faça backoff.
    InternalError:
      description: Erro interno (transitório).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "internal error" }
  schemas:
    Error:
      type: object
      description: Envelope de erro uniforme (não é RFC 7807).
      required: [error]
      properties:
        error: { type: string, description: Mensagem curta e segura (sem detalhe interno). }
...
    CreateClientRequest:
      type: object
      description: >-
        Provisionamento de empresa-cliente (modelo b). NÃO aceita `account_id`:
        a Conta vem sempre da chave (server-side, imutável — A01/T6).
      properties:
        name: { type: string, description: Nome/label da empresa-cliente (opcional). }
    Client:
      type: object
      description: Empresa-cliente provisionada, já vinculada à Conta da chave.
      properties:
        tenant_id:
          type: string
          description: Identificador da empresa-cliente. Use-o no header `X-Client-Tenant`.
        account_id: { type: string, description: Conta dona (derivada da chave). }
        name: { type: string }
...